[改良] プロヴィンスの労働力で小数点以下第1位を許容するよう変更 fixed #133

+++ HoI2Editor/Models/Province.cs
@@ -95,7 +95,7 @@ namespace HoI2Editor.Models
         /// <summary>
         ///     労働力
         /// </summary>
-        public int Manpower { get; set; }
+        public double Manpower { get; set; }
 
         /// <summary>
         ///     石油
